Combine the concepts/ and working/ docs into reference/.
The previous "Workflow Development" section represents a complete and
detailed reference documentation for all the covered topics, which is
now better reflected by the name.

Furthermore, both concepts and practical implementation are tightly
related and in an effort to reduce the overall number of sections, all
concept sub-sections have been integrated with the reference
documentation sections.